requirements.txt
streamlit
yfinance
plotly
google-genai